+++ js/big-pictures.js
@@ -1,11 +1,60 @@
 import { isEscapeKey } from './util.js';
-import { renderComments, initCommentList } from './comment.js';
+
+const COMMENTS_COUNT_TO_SHOW = 5;
+
 
 const bigPictureElement = document.querySelector('.big-picture');
 const bodyElement = document.querySelector('body');
 const closeButtonElement = bigPictureElement.querySelector('.big-picture__cancel');
 
+const commentElement = bigPictureElement.querySelector('.social__comment');
+const commentsListElement = bigPictureElement.querySelector('.social__comments');
+const commentCountElement = bigPictureElement.querySelector('.social__comment-shown-count');
+const totalCommentCountElement = bigPictureElement.querySelector('.social__comment-total-count');
+const commentsLoaderElement = bigPictureElement.querySelector('.comments-loader');
+
+let commentsCountShown = 0;
+let comments = [];
+
+const createComment = function ({ avatar, message, name }) {
+  const newComment = commentElement.cloneNode(true);
+
+  newComment.querySelector('.social__picture').src = avatar;
+  newComment.querySelector('.social__picture').alt = name;
+  newComment.querySelector('.social__text').textContent = message;
+
+  return newComment;
+};
+
+const renderComments = function () {
+  commentsCountShown += COMMENTS_COUNT_TO_SHOW;
+
+  if (commentsCountShown >= comments.length) {
+    commentsLoaderElement.classList.add('hidden');
+    commentsCountShown = comments.length;
+  } else {
+    commentsLoaderElement.classList.remove('hidden');
+  }
+
+  const fragment = document.createDocumentFragment();
+  for (let i = 0; i < commentsCountShown; i++) {
+    const comment = createComment(comments[i]);
+    fragment.append(comment);
+  }
+
+  commentsListElement.innerHTML = '';
+  commentsListElement.append(fragment);
+
+  commentCountElement.textContent = commentsCountShown;
+  totalCommentCountElement.textContent = comments.length;
+};
+
+const onCommentsLoaderClick = function () {
+  renderComments();
+};
+
 const hidePicture = function () {
+  commentsCountShown = 0;
   bigPictureElement.classList.add('hidden');
   bodyElement.classList.remove('modal-open');
 
@@ -37,12 +86,17 @@ const showPicture = function (pictureData) {
       hidePicture();
     }
   });
-  renderComments(pictureData.comments);
-  initCommentList();
+
+  comments = pictureData.comments;
+  if (comments.length > 0) {
+    renderComments();
+  }
 
   renderPicture(pictureData);
 };
 
 closeButtonElement.addEventListener('click', onCloseButtonClick);
+
+commentsLoaderElement.addEventListener('click', onCommentsLoaderClick);
 
 export { showPicture };
